feat: add get movies with showtimes

+++ src/movies/movies.controller.ts
@@ -10,7 +10,6 @@ import {
   Patch,
   Post,
   Query,
-  UnauthorizedException,
   UseGuards,
 } from '@nestjs/common';
 import { CreateMovieDto } from './dto/create-movie.dto';
@@ -22,75 +21,81 @@ import { AuthGuard } from '@nestjs/passport';
 import { GetUser } from '../auth/get-user.decorator';
 import { User } from '../users/user.entity';
 import { UserRole } from '../users/user-role.enum';
+import { Cinema } from 'src/cinemas/cinema.entity';
 
 @Controller('movies')
 export class MoviesController {
   private logger = new Logger('MoviesController');
 
+  private allowedRole: Array<UserRole> = [UserRole.ADMIN, UserRole.SUPER_ADMIN];
+
   constructor(private moviesService: MoviesService) {}
 
   @Get()
-  getMovies(@Query() filterDto: GetMoviesFilterDto): Promise<Array<Movie>> {
+  async getMovies(
+    @Query() filterDto: GetMoviesFilterDto,
+  ): Promise<Array<Movie>> {
     return this.moviesService.getMovies(filterDto);
   }
 
   @Get('/:id')
-  getMovieById(@Param('id', ParseIntPipe) id: number): Promise<Movie> {
+  async getMovieById(@Param('id', ParseIntPipe) id: number): Promise<Movie> {
     return this.moviesService.getMovieById(id);
   }
 
+  @Get('/cinemas/:id')
+  async getCinemasFromMovieId(
+    @Param('id', ParseIntPipe) id: number,
+  ): Promise<Cinema[]> {
+    return this.moviesService.getCinemasFromMovieId(id);
+  }
+
   @Post()
-  // @UseGuards(AuthGuard())
+  @UseGuards(AuthGuard())
   createMovie(
     @Body() createMovieDto: CreateMovieDto,
-    // @GetUser() user: User,
+    @GetUser() user: User,
   ): Promise<Movie> {
-    // const allowedRole: Array<UserRole> = [UserRole.ADMIN, UserRole.SUPER_ADMIN];
+    if (!this.allowedRole.includes(user.role)) {
+      throw new ForbiddenException();
+    }
 
-    // if (!allowedRole.includes(user.role)) {
-    //   throw new ForbiddenException();
-    // }
-
-    // this.logger.verbose(
-    //   `User "${user.username}" creating a new movie. Data: ${JSON.stringify(
-    //     createMovieDto,
-    //   )}`,
-    // );
+    this.logger.verbose(
+      `User "${user.username}" creating a new movie. Data: ${JSON.stringify(
+        createMovieDto,
+      )}`,
+    );
 
     return this.moviesService.createMovie(createMovieDto);
   }
 
   @Delete('/:id')
-  // @UseGuards(AuthGuard())
+  @UseGuards(AuthGuard())
   deleteMovie(
     @Param('id', ParseIntPipe) id: number,
-    // @GetUser() user: User,
+    @GetUser() user: User,
   ): Promise<void> {
-    // const allowedRole: Array<UserRole> = [UserRole.ADMIN, UserRole.SUPER_ADMIN];
+    if (!this.allowedRole.includes(user.role)) {
+      throw new ForbiddenException();
+    }
 
-    // if (!allowedRole.includes(user.role)) {
-    //   throw new ForbiddenException();
-    // }
-
-    // this.logger.verbose(
-    //   `User "${user.username}" deleting a movie with ID ${id}.`,
-    // );
+    this.logger.verbose(
+      `User "${user.username}" deleting a movie with ID ${id}.`,
+    );
 
     return this.moviesService.deleteMovie(id);
   }
 
   @Patch('/:id')
-  // @UseGuards(AuthGuard())
+  @UseGuards(AuthGuard())
   updateMovie(
     @Param('id', ParseIntPipe) id: number,
     @Body() updateMovieDto: UpdateMovieDto,
-    // @GetUser() user: User,
+    @GetUser() user: User,
   ): Promise<Movie> {
-    // const allowedRole: Array<UserRole> = [UserRole.ADMIN, UserRole.SUPER_ADMIN];
-
-    // if (!allowedRole.includes(user.role)) {
-    //   throw new ForbiddenException();
-    // }
+    if (!this.allowedRole.includes(user.role)) {
+      throw new ForbiddenException();
+    }
 
     return this.moviesService.updateMovie(id, updateMovieDto);
   }
